Split horizon into two other modules
