+++ clients/vault/src/horizon.rs
@@ -1,8 +1,5 @@
-use crate::error::Error;
 use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
-use runtime::{Balance, SpacewalkParachain, RedeemEvent, SpacewalkPallet};
 use serde::{Deserialize, Deserializer};
-use service::{spawn_cancelable, Error as ServiceError, ShutdownSender};
 use sp_core::ed25519;
 use sp_runtime::{
     scale_info::TypeInfo,
@@ -17,17 +14,11 @@ use sp_std::{
     str::from_utf8,
     vec::Vec,
 };
-use std::time::Duration;
 use stellar::{
-    network::TEST_NETWORK,
     types::{AssetAlphaNum12, AssetAlphaNum4},
-    Asset, PublicKey, SecretKey,
+    Asset, PublicKey,
 };
 use substrate_stellar_sdk as stellar;
-use tokio::time::sleep;
-
-const POLL_INTERVAL: u64 = 5000;
-const SUBMISSION_TIMEOUT_PERIOD: u64 = 10000;
 
 // This represents each record for a transaction in the Horizon API response
 #[derive(Deserialize, Encode, Decode, Default, Debug)]
@@ -306,320 +297,4 @@ impl Convert<(Vec<u8>, Vec<u8>), Result<CurrencyId, ()>> for StringCurrencyConve
         let asset_code = from_utf8(a.0.as_slice()).map_err(|_| ())?;
         (asset_code, public_key.into_binary()).try_into().map_err(|_| ())
     }
-}
-
-/////////////////
-/// Deposit
-
-fn is_escrow(escrow_key: &String, public_key: [u8; 32]) -> bool {
-    let escrow_keypair: SecretKey = SecretKey::from_encoding(escrow_key).unwrap();
-    return public_key == *escrow_keypair.get_public().as_binary();
-}
-
-/// Fetch recent transactions from remote and deserialize to HorizonResponse
-/// Since the limit in the request url is set to one it will always fetch just one
-async fn fetch_latest_txs(escrow_secret_key: &String) -> Result<HorizonTransactionsResponse, Error> {
-    let escrow_keypair: SecretKey = SecretKey::from_encoding(escrow_secret_key).unwrap();
-    let escrow_address = escrow_keypair.get_public();
-
-    let request_url = String::from("https://horizon-testnet.stellar.org/accounts/")
-        + str::from_utf8(escrow_address.to_encoding().as_slice()).map_err(|_| Error::HttpFetchingError)?
-        + "/transactions?order=desc&limit=1";
-
-    let horizon_response = reqwest::get(request_url.as_str())
-        .await
-        .map_err(|_| Error::HttpFetchingError)?
-        .json::<HorizonTransactionsResponse>()
-        .await
-        .map_err(|_| Error::HttpFetchingError)?;
-
-    Ok(horizon_response)
-}
-
-static mut LAST_TX_ID: Option<Vec<u8>> = None;
-
-fn is_unhandled_transaction(tx: &Transaction) -> bool {
-    const UP_TO_DATE: () = ();
-    let latest_tx_id_utf8 = &tx.id;
-
-    unsafe {
-        let prev_tx_id = &LAST_TX_ID;
-        let initial = !matches!(prev_tx_id, Some(_));
-
-        let result = match prev_tx_id {
-            Some(prev_tx_id) => {
-                if prev_tx_id == latest_tx_id_utf8 {
-                    Err(UP_TO_DATE)
-                } else {
-                    Ok(latest_tx_id_utf8.clone())
-                }
-            }
-            None => Ok(latest_tx_id_utf8.clone()),
-        };
-
-        match result {
-            Ok(latest_tx_id) => {
-                LAST_TX_ID = Some(latest_tx_id.clone());
-                if !initial {
-                    tracing::info!(
-                        "Found new transaction from Horizon (id {:#?}). Starting to process new transaction",
-                        str::from_utf8(&latest_tx_id).unwrap()
-                    );
-
-                    true
-                } else {
-                    tracing::info!("Initial transaction handled");
-                    false
-                }
-            }
-            Err(UP_TO_DATE) => {
-                tracing::info!("Already up to date");
-                false
-            }
-        }
-    }
-}
-
-async fn report_transaction(parachain_rpc: &SpacewalkParachain, tx: &Transaction) -> Result<(), Error> {
-    // Decode transaction to Base64
-    let tx_xdr = base64::decode(&tx.envelope_xdr).unwrap();
-
-    // Send new transaction to spacewalk bridge pallet
-    let result = parachain_rpc.report_stellar_transaction(&tx_xdr).await;
-    match result {
-        Ok(_) => Ok(()),
-        Err(e) => Err(Error::RuntimeError(e)),
-    }
-}
-
-async fn fetch_horizon_and_process_new_transactions(parachain_rpc: &SpacewalkParachain, escrow_secret_key: &String) {
-    let res = fetch_latest_txs(escrow_secret_key).await;
-    let transactions = match res {
-        Ok(txs) => txs._embedded.records,
-        Err(e) => {
-            tracing::warn!("Failed to fetch transactions: {:?}", e);
-            return;
-        }
-    };
-
-    if transactions.len() > 0 {
-        let tx = &transactions[0];
-        if is_unhandled_transaction(tx) {
-            let result = report_transaction(parachain_rpc, tx).await;
-            match result {
-                Ok(_) => {
-                    tracing::info!("Reported Stellar transaction to spacewalk pallet");
-                }
-                Err(e) => {
-                    tracing::warn!("Failed to process transaction: {:?}", e);
-                }
-            }
-        }
-    }
-}
-
-pub async fn poll_horizon_for_new_transactions(
-    parachain_rpc: SpacewalkParachain,
-    escrow_secret_key: String,
-) -> Result<(), ServiceError> {
-    // Start polling horizon every 5 seconds
-    loop {
-        fetch_horizon_and_process_new_transactions(&parachain_rpc, &escrow_secret_key).await;
-
-        sleep(Duration::from_millis(POLL_INTERVAL)).await;
-    }
-}
-
-/////////////////
-/// Withdrawal
-
-async fn fetch_from_remote(request_url: &str) -> Result<HorizonAccountResponse, Error> {
-    tracing::info!("Sending request to: {}", request_url);
-
-    let response = reqwest::get(request_url)
-        .await
-        .map_err(|_| Error::HttpFetchingError)?
-        .json::<HorizonAccountResponse>()
-        .await
-        .map_err(|_| Error::HttpFetchingError)?;
-
-    Ok(response)
-}
-
-async fn fetch_latest_seq_no(stellar_addr: &str) -> Result<u64, Error> {
-    let request_url = String::from("https://horizon-testnet.stellar.org/accounts/") + stellar_addr;
-
-    let horizon_response = fetch_from_remote(request_url.as_str()).await.map_err(|e| {
-        tracing::error!("fetch_latest_seq_no error: {:?}", e);
-        Error::HttpFetchingError
-    })?;
-
-    String::from_utf8(horizon_response.sequence)
-        .map(|string| string.parse::<u64>().unwrap())
-        .map_err(|_| Error::SeqNoParsingError)
-}
-
-fn sign_stellar_tx(
-    tx: stellar::types::Transaction,
-    secret_key: SecretKey,
-) -> Result<stellar::TransactionEnvelope, Error> {
-    let mut envelope = tx.into_transaction_envelope();
-    envelope.sign(&TEST_NETWORK, vec![&secret_key])?;
-
-    Ok(envelope)
-}
-
-fn submit_stellar_tx(tx: stellar::TransactionEnvelope) -> Result<(), Error> {
-    let mut last_error: Option<Error> = None;
-
-    for attempt in 1..=3 {
-        tracing::debug!("Attempt #{} to submit Stellar transaction…", attempt);
-
-        match try_once_submit_stellar_tx(&tx) {
-            Ok(result) => {
-                return Ok(result);
-            }
-            Err(error) => {
-                last_error = Some(error);
-            }
-        }
-    }
-
-    // Can only panic if no submission was ever attempted
-    Err(last_error.unwrap())
-}
-
-fn try_once_submit_stellar_tx(tx: &stellar::TransactionEnvelope) -> Result<(), Error> {
-    let horizon_base_url = "https://horizon-testnet.stellar.org";
-    let horizon = stellar::horizon::Horizon::new(horizon_base_url);
-
-    tracing::info!("Submitting transaction to Stellar network: {}", horizon_base_url);
-
-    let _response = horizon
-        .submit_transaction(&tx, SUBMISSION_TIMEOUT_PERIOD, true)
-        .map_err(|error| {
-            match error {
-                stellar::horizon::FetchError::UnexpectedResponseStatus { status, body } => {
-                    tracing::error!("Unexpected HTTP request status code: {}", status);
-                    tracing::error!("Response body: {}", str::from_utf8(&body).unwrap());
-                }
-                _ => (),
-            }
-            Error::HttpFetchingError
-        })?;
-
-    Ok(())
-}
-
-fn create_withdrawal_tx(
-    escrow_secret_key: &String,
-    stellar_addr: &stellar::PublicKey,
-    seq_num: i64,
-    asset: stellar::Asset,
-    amount: Balance,
-) -> Result<stellar::Transaction, Error> {
-    let destination_addr = stellar_addr.as_binary();
-
-    let source_keypair: SecretKey = SecretKey::from_encoding(escrow_secret_key).unwrap();
-
-    let source_pubkey = source_keypair.get_public().clone();
-
-    let mut tx = stellar::Transaction::new(source_pubkey, seq_num, Some(10_000), None, None)?;
-
-    tx.append_operation(stellar::Operation::new_payment(
-        stellar::MuxedAccount::KeyTypeEd25519(*destination_addr),
-        asset,
-        stellar::StroopAmount(BalanceConversion::lookup(amount).map_err(|_| Error::BalanceConversionError)?),
-    )?)?;
-
-    Ok(tx)
-}
-
-async fn execute_withdrawal(
-    escrow_secret_key: &String,
-    amount: Balance,
-    currency_id: CurrencyId,
-    destination_stellar_address: PublicKey,
-) -> Result<(), Error> {
-    let asset = CurrencyConversion::lookup(currency_id)?;
-
-    let escrow_keypair: SecretKey = SecretKey::from_encoding(escrow_secret_key).unwrap();
-    let escrow_encoded = escrow_keypair.get_public().to_encoding().clone();
-    let escrow_address = str::from_utf8(escrow_encoded.as_slice())?;
-
-    tracing::info!("Execute withdrawal: ({:?}, {:?})", currency_id, amount,);
-
-    let seq_no = fetch_latest_seq_no(escrow_address).await.map(|seq_no| seq_no + 1)?;
-    let transaction = create_withdrawal_tx(
-        escrow_secret_key,
-        &destination_stellar_address,
-        seq_no as i64,
-        asset,
-        amount,
-    )?;
-    let signed_envelope = sign_stellar_tx(transaction, escrow_keypair)?;
-
-    let result = submit_stellar_tx(signed_envelope);
-    tracing::info!(
-        "✔️  Successfully submitted withdrawal transaction to Stellar, crediting {}",
-        str::from_utf8(destination_stellar_address.to_encoding().as_slice()).unwrap()
-    );
-
-    result
-}
-
-pub async fn listen_for_redeem_requests(
-    shutdown_tx: ShutdownSender,
-    parachain_rpc: SpacewalkParachain,
-    escrow_secret_key: String,
-) -> Result<(), ServiceError> {
-    tracing::info!("Starting to listen for redeem requests…");
-    parachain_rpc
-        .on_event::<RedeemEvent, _, _, _>(
-            |event| async {
-                tracing::info!("Received redeem request: {:?}", event);
-
-                // within this event callback, we captured the arguments of listen_for_redeem_requests
-                // by reference. Since spawn requires static lifetimes, we will need to capture the
-                // arguments by value rather than by reference, so clone these:
-                let secret_key = escrow_secret_key.clone();
-                let asset_code = event.asset_code.clone();
-                let asset_issuer = event.asset_issuer.clone();
-                let stellar_user_id = event.stellar_user_id.clone();
-                let stellar_vault_id = event.stellar_vault_id.clone();
-                let amount = event.amount.clone();
-
-                if !is_escrow(&escrow_secret_key, stellar_vault_id.try_into().unwrap()) {
-                    tracing::info!("Rejecting redeem request: not an escrow");
-                    return;
-                }
-
-                // Spawn a new task so that we handle these events concurrently
-                spawn_cancelable(shutdown_tx.subscribe(), async move {
-                    tracing::info!("Executing redeem #{:?}", event);
-
-                    let currency_id = StringCurrencyConversion::convert((asset_code, asset_issuer)).unwrap();
-                    let destination_stellar_address = stellar::PublicKey::from_encoding(stellar_user_id).unwrap();
-
-                    let result =
-                        execute_withdrawal(&secret_key, amount, currency_id, destination_stellar_address).await;
-
-                    match result {
-                        Ok(_) => tracing::info!(
-                            "Completed redeem request with amount {}",
-                            // event.redeem_id,
-                            event.amount
-                        ),
-                        Err(e) => tracing::error!(
-                            "Failed to process redeem request: {}",
-                            // event.redeem_id,
-                            e.to_string()
-                        ),
-                    }
-                });
-            },
-            |error| tracing::error!("Error reading redeem event: {}", error.to_string()),
-        )
-        .await?;
-    Ok(())
 }
